rekomendasi 2.0

+++ Logic/deleteItem.php
@@ -3,14 +3,21 @@
 include_once "../Auth/connection.php";
 session_start();
 
-// $admin_id = $_SESSION['Type'];
-// if($admin_id !== 1){
-//     header('Location: Logout.php');
-// }
+$_SESSION['csrf_token'] = bin2hex(random_bytes(32));
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['delete'])) {
     $id = $_POST['delete'];
-    
+
+    if(!hash_equals($_SESSION['csrf_token'], $_POST['csrf_token'])){
+        echo "Token csrf tidak valid";
+        exit();
+    }   
+    $id = filter_var($_POST['delete'], FILTER_VALIDATE_INT);
+    if($id == false){
+        echo "ID tidak valid";
+        exit();
+    }
+
     $query = "DELETE FROM menu WHERE id = ?";
     $stmt = $conn->prepare($query);
     $stmt->bind_param("i", $id);
